llamaindex_api/requirements.txt
llama-index==0.12.12
fastapi==0.115.8
python-dotenv==1.0.1
llama-index-llms-openrouter==0.3.1
llama-index-embeddings-huggingface==0.5.1